Fix broken phase field test
I don't understand the error very well. I've fixed it by once
again switching between lax.cond and np.where, which must be the
sixth time I've done this in the plasticity models. I'm still not
sure what the rules are for short circuiting in jax and how to be
certain that the false branch of the code won't be executed under
AD and/or jit. Comprehensive unit testing is in order.