Fix bug in conversion of linear mesh to higher order
Also added a test to catch this bug.

It doesn't always happen in practice. When adding the higher order
nodes on edges, the condition that checks for the presence of a
second (right-hand) neighboring triangle was wrong, because it
checked for an ordinal > 0, but 0 is a valid element ordinal.
The sentinel value for "no existing right-side neighbor" is -1,
so the conditional should have checked for an ordinal >= 0.

The bug only happened if element 0 was the right-side neighbor of
any edge. If all the global edges that touched element 0 happened
to go around it in the positive sense, this bug wouldn't happen.
This is how it avoided the tests until now.