Implement missing functions needed to use multi-material simulations with internal vars
When materials have different numbers of internal variables, this
pads the array so that each quadrature point has enough space for
the largest number of internal variables. This is not
memory efficient, but it works for now.

The while multi-material paradigm really needs a review and
probably a refactoring (if not new implementation).

Maybe "blocks" should be another array dimension.