Rewrite dynamics as a continuous lagrangian field problem
Before: We precomputed the element mass matrices and used these
in the dynamics functions directly. The kinetic enerrgy was
thus computed by quadratic matrix forms.

Now: Write the Newmark corrector phase as a continuous field
problem. The action of the mass matrix is now implicit - it's
in the kinetic energy part of the Lagrangian density. You
can think of this as fusing the kernel.

* Makes the dynamics formulation more uniform with statics.
* Exposes further opportunities to reuse functions for both
  statics and dynamics. The sparse Hessian assembly, for instance,
  is the same, just the lagrangian density function fed to it
  changes.
* Materials now must have a density member variable. A default
  value of 'None' is created if the user doesn't specify it.
  This will allow statics calculations to be done without
  requiring the density to be specified. Dynamics, which needs
  density, will emit a clear error if the user forgets to
  specify it.
* Makes it easy to get the action of the mass matrix on a
  vector - just take the gradient of the kinetic energy function
  at the value of the vector. This is also efficient.