Fix bug made when I fused the kernel
Can't be done with a fused kernel, since the kinetic energy and
strain energy terms take different fields in the lagrangian
densite. The kinetic energy takes the algorithmic velocity, and the
strain energy takes the displacement.

I caught this with the dynamic uniaxial app. We need a dynamic
patch test so this thing is caught in the tests.