Create new classes so that a quadrature rule isn't needed before mesh is read
I'm trying to accomplish two things:

1. Compute the master element shape functions and derivatives only
   once, so that construction of the function space can happen
   inside jitted code. This is to allow shape sensitivities. We want
   gradients of QOIs with respect to nodal coordinates. Only the
   mapping of the shape function derivatives from reference domain
   to physical domain needs to be traced by Jax.
2. Defer the instantiation of the quadrature rule until after the
   mesh is generated (or read in). This is because we may want to
   create several quadrature rules on the same mesh for different
   QOIs.

The first point suggests that a data structure be created that
holds the master element shape functions once and for all. However,
point 2 demands that this data structure not be the same as the one
that holds the nodal point coordinates in the master element, as
this is needed at mesh construction time.

The compromise here is to create two types: NodalBasis, which
just holds the nodal coordinates in the master element, as well
as which nodes are on vertices and faces, etc; MasterElement,
which holds all this data plus the shape functions and gradients
at a set of quadrature points.