Test in a better way to avoid expected failure
Test conditions remain the same. Before, to make sure there were
no warnings about inconsistent data sizes, we asserted that a
warning was raised and then marked the test as an expected failure.
This is not really what the expected failure decorator is for,
and is alarming to see on the automated test output.

Now, we catch warnings on the same tests, and promote them to
errors with a context manager. This way, test FAILURE means the
warning was generated.