Make Neo-Hookean consistent with small strains
For the "coupled" version of the Neo-Hookean model, the material
parameters were misleading. The property labeled as the bulk
modulus was used in a place where it really acted as the
Lame's lambda parameter. The model wasn't wrong per se, but it
would not match infinitesimal strains in the expected way.

This commit adds the Lame lambda parameter and switches to the
standard usage.