Remove unneccessary infinities for negative EQPS in hardening functions
We have a solver that has a lower bound on the EQPS increment,
(which is set to 0) so we're already protected against invalid
values. When the hardening functions return infinity on negative
EQPS, this can get triggered during iterations due to floating
point cancellation or during an iterate that is going to be
safely changed by the solver, which is counterproductive.

Also, the only hardening function with a singular point (power law)
will automatically return a NaN or inf when the singularity is hit,
so there's no need to put an external condition.

Also in this commit:
* Remove wildcard import
* Fix incorrect exception type name from standard library (should
  be "ValueError", not "valueError"