Initialize internal variables in a simpler way
Before, the mechanics module asked the material for copies of the
initial values of the internal variables, telling it how many
elements and quadrature points per element there were.

Now, the material gives back initial values for the internal
variables at a single material point, and the mechanics module
makes as many copies as needed. This is a better division of
responsibility, avoiding passing finite element model data
to a material model.