Fix higher order derivatives of matrix functions
Before: the primal output was computed in-line in the custom jvp
function. The advantage is that this avoids a second call to the
eigendecomposition. The downside is that this in-line computation
doesn't itself have a custom jvp, so its derivative can be wrong.

After: I re-compute the primal value through the base function
(e.g., log_symm), which has the custom jvp defined on it. The
eigendecomposition is repeated. We can refactor to eliminate this
later if profiling reveals it to be a performance bottleneck.