Add a higher order quadrature rule (exact up to p=10)
Also: now a ValueError is emitted if a user asks for a quadrature
rule precision that is above what we have implemented. Before
the error was a bit obscure.

Improved the unit testing of the quadrature rules by making
subtests for every polynomial degree check. This way, if the test
fails, you can figure out which quadrature rules are broken.