Fix GMRES solver errors in tests triggered in new scipy version
Scipy GMRES now modifies the preconditioned residual in place. This
is perfectly reasonable, but we were passing in jax numpy arrays,
which do not allow this. This was unsafe, but we got away with it.

This commit makes deep copies of jax arrays to plain numpy arrays
in the operators and initial data that the scipy gmres sees.